crawler/requirements.txt
beautifulsoup4==4.12.2
flask==2.3.3
psycopg2-binary==2.9.7
requests==2.31.0
